requirements-test.txt
django>=1.9.4
coverage
mock>=1.0.1
flake8>=2.1.0
factory_boy
